+++ src/setter_computed_/index.js
@@ -1,4 +1,5 @@
-import { atom } from 'nanostores'
+import { run } from '@ctx-core/function'
+import { atom, onMount, onStop } from 'nanostores'
 import { computed_ } from '../computed_/index.js'
 /** @type {import('./index.d.ts').setter_computed__T} */
 export const setter_computed_ = (stores, cb)=>{
@@ -11,29 +12,30 @@ export const setter_computed_ = (stores, cb)=>{
 				/** @type {import('nanostores').AnyStore[]} */
 				store_val_a)
 	const return_payload_atom = atom()
-	let _store_val_a = []
-	let run = store_val_a=>
-		cb(
+	let cb__run = store_val_a=>{
+		return cb(
 			isArray
 			? store_val_a
+				|| store_a.map(()=>undefined)
 			: store_val_a && store_val_a[0],
 			val=>{
-				if (
-					store_val_a
-					&& store_val_a.every((store_val, $i)=>
-						store_val === _store_val_a[$i])
-				) {
-					return_payload_atom.set(val)
-				} else {
-					console.warn('setter_computed_|run|stale inputs', cb)
-				}
+				return_payload_atom.set(val)
 			})
-	store_val_a_.subscribe($=>{
-		_store_val_a = $
-		run($)
+	}
+	let unsubscribe
+	/** @type {import('./index.d.ts').SetterComputedAtom_} */
+	const setter_computed =
+		computed_(
+			return_payload_atom,
+			val=>val)
+	onMount(setter_computed, ()=>{
+		unsubscribe = store_val_a_.subscribe($=>{
+			cb__run($)
+		})
 	})
-	return (
-		/** @type {import('./index.d.ts').SetterComputedAtom_} */
-		computed_(return_payload_atom, val=>val))
+	onStop(
+		setter_computed,
+		()=>run(unsubscribe))
+	return setter_computed
 }
 export { setter_computed_ as setter_computed$ }